refactor: remove active_citizens column (#51)
