+++ data-sources/citizens/generate-citizens-csv.ts
@@ -1,6 +1,6 @@
 const Web3 = require('web3')
-const Passport = require('./abis/Passport.json')
-const VotingEscrow = require('./abis/VotingEscrow.json')
+const Passport = require('../abis/Passport.json')
+const VotingEscrow = require('../abis/VotingEscrow.json')
 const csvWriter = require('csv-writer')
 const fs = require('fs')
 const ethers = require('ethers')
