+++ data-sources/citizens/output/citizen-count-per-week.csv
@@ -1,29 +1,29 @@
-week_end,total_citizens,active_citizens
-2022-06-05,156,0
-2022-06-12,158,0
-2022-06-19,161,0
-2022-06-26,164,0
-2022-07-03,168,0
-2022-07-10,168,0
-2022-07-17,168,0
-2022-07-24,171,0
-2022-07-31,173,0
-2022-08-07,174,0
-2022-08-14,175,0
-2022-08-21,176,0
-2022-08-28,176,0
-2022-09-04,177,0
-2022-09-11,178,0
-2022-09-18,179,0
-2022-09-25,179,0
-2022-10-02,179,0
-2022-10-09,182,0
-2022-10-16,183,0
-2022-10-23,185,0
-2022-10-30,185,0
-2022-11-06,189,0
-2022-11-13,222,0
-2022-11-20,232,0
-2022-11-27,234,0
-2022-12-04,236,0
-2022-12-11,241,0
+week_end,total_citizens
+2022-06-05,156
+2022-06-12,158
+2022-06-19,161
+2022-06-26,164
+2022-07-03,168
+2022-07-10,168
+2022-07-17,168
+2022-07-24,171
+2022-07-31,173
+2022-08-07,174
+2022-08-14,175
+2022-08-21,176
+2022-08-28,176
+2022-09-04,177
+2022-09-11,178
+2022-09-18,179
+2022-09-25,179
+2022-10-02,179
+2022-10-09,182
+2022-10-16,183
+2022-10-23,185
+2022-10-30,185
+2022-11-06,189
+2022-11-13,222
+2022-11-20,232
+2022-11-27,234
+2022-12-04,236
+2022-12-11,241
